backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
openai
beautifulsoup4
supabase 